+++ api/src/queries/getSongsPlayedPerDay.sql
@@ -6,7 +6,7 @@ WITH daily_counts AS (
         s.album,
         s."image",
         s.sync_id,
-        COUNT(*) AS listens
+        COUNT(*)::INT AS listens
     FROM activities a
     JOIN songs s 
       ON a.song = s.name
